Bugfix for invalid strings in PHP 8 - he_IL
sprintf expects placeholders to use the format `%1$s`.
Prior to v8, placeholders with a space would be interpreted incorrectly, but would not throw an error.
In v8 PHP throws a fatal error, causing all the Gibbons to escape.
Live example: https://onlinephp.io/c/523c3
This commit removes the extraneous spacing on 361 placeholders.